requirements.txt
flask>=2.0.1
flask-login>=0.5.0
pyyaml>=5.4.1
markdown2>=2.4.1
requests>=2.20.0